+++ modules/main/src/main/java/com/nisovin/shopkeepers/util/inventory/EnchantmentUtils.java
@@ -6,13 +6,13 @@ import java.util.Map;
 
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
-import org.bukkit.Registry;
 import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.EnchantmentStorageMeta;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import com.nisovin.shopkeepers.api.internal.util.Unsafe;
+import com.nisovin.shopkeepers.compat.Compat;
 import com.nisovin.shopkeepers.util.bukkit.NamespacedKeyUtils;
 import com.nisovin.shopkeepers.util.java.ConversionUtils;
 import com.nisovin.shopkeepers.util.java.Validate;
@@ -68,7 +68,7 @@ public final class EnchantmentUtils {
 		}
 
 		// Lookup by key:
-		return Registry.ENCHANTMENT.get(namespacedKey);
+		return Compat.getProvider().getRegistry(Enchantment.class).get(namespacedKey);
 	}
 
 	/**