+++ modules/v1_21_R8_paper/src/main/java/com/nisovin/shopkeepers/compat/v1_21_R8_paper/CompatProviderImpl.java
@@ -1,10 +1,12 @@
 package com.nisovin.shopkeepers.compat.v1_21_R8_paper;
 
 import java.lang.reflect.Field;
-import java.lang.reflect.Method;
+import java.lang.reflect.ParameterizedType;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.bukkit.Bukkit;
-import org.bukkit.ExplosionResult;
+import org.bukkit.Keyed;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.Registry;
@@ -19,22 +21,16 @@ import org.bukkit.craftbukkit.inventory.CraftItemStack;
 import org.bukkit.craftbukkit.inventory.CraftMerchant;
 import org.bukkit.craftbukkit.util.CraftMagicNumbers;
 import org.bukkit.entity.AbstractVillager;
-import org.bukkit.entity.Chicken;
 import org.bukkit.entity.CopperGolem;
 import org.bukkit.entity.CopperGolem.Oxidizing;
-import org.bukkit.entity.Cow;
 import org.bukkit.entity.Entity;
 import org.bukkit.entity.Golem;
 import org.bukkit.entity.LivingEntity;
 import org.bukkit.entity.Mannequin;
 import org.bukkit.entity.Mob;
-import org.bukkit.entity.Pig;
 import org.bukkit.entity.Player;
 import org.bukkit.entity.Pose;
-import org.bukkit.entity.Salmon;
 import org.bukkit.entity.Villager;
-import org.bukkit.event.block.BlockExplodeEvent;
-import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.MainHand;
@@ -50,7 +46,7 @@ import com.nisovin.shopkeepers.api.internal.util.Unsafe;
 import com.nisovin.shopkeepers.compat.CompatProvider;
 import com.nisovin.shopkeepers.shopobjects.living.LivingEntityAI;
 import com.nisovin.shopkeepers.util.annotations.ReadOnly;
-import com.nisovin.shopkeepers.util.bukkit.RegistryUtils;
+import com.nisovin.shopkeepers.util.bukkit.ServerUtils;
 import com.nisovin.shopkeepers.util.data.container.DataContainer;
 import com.nisovin.shopkeepers.util.inventory.ItemStackComponentsData;
 import com.nisovin.shopkeepers.util.inventory.ItemStackMetaTag;
@@ -82,20 +78,43 @@ import net.minecraft.world.item.trading.MerchantOffers;
 
 public final class CompatProviderImpl implements CompatProvider {
 
+	private static final Map<Class<?>, RegistryKey<?>> CLASS_TO_REGISTRY_KEY = new HashMap<>();
+
+	static {
+		try {
+			for (var field : RegistryKey.class.getFields()) {
+				if (field.getType() != RegistryKey.class) {
+					continue;
+				}
+
+				// Get the type from the RegistryKey generic parameter on the field:
+				var fieldType = (ParameterizedType) field.getGenericType();
+				var typeArgument = fieldType.getActualTypeArguments()[0];
+				Class<?> registryClass;
+				if (typeArgument instanceof Class<?> typeArgumentClass) {
+					registryClass = typeArgumentClass;
+				} else if (typeArgument instanceof ParameterizedType typeArgumentParameterized) {
+					registryClass = Unsafe.castNonNull(typeArgumentParameterized.getRawType());
+				} else {
+					throw new RuntimeException("Unexpected RegistryKey type parameter for field: "
+							+ field.getName());
+				}
+
+				RegistryKey<?> registryKey = Unsafe.castNonNull(field.get(null));
+				CLASS_TO_REGISTRY_KEY.put(registryClass, registryKey);
+			}
+		} catch (ReflectiveOperationException ex) {
+			throw new RuntimeException(ex);
+		}
+	}
+
 	private final TagParser<Tag> tagParser = Unsafe.castNonNull(TagParser.create(NbtOps.INSTANCE));
 
 	private final Field craftItemStackHandleField;
-	private final Method cowSetVariantMethod;
 
 	public CompatProviderImpl() throws Exception {
 		craftItemStackHandleField = CraftItemStack.class.getDeclaredField("handle");
 		craftItemStackHandleField.setAccessible(true);
-
-		// TODO Spigot 1.21.5 remaps "Cow" to "AbstractCow" unless api-version <= "1.21.5". So we
-		// use reflection here to support the new cow variants anyway while still supporting older
-		// api versions.
-		var cowClass = Class.forName("org.bukkit.entity.Cow");
-		cowSetVariantMethod = cowClass.getMethod("setVariant", Cow.Variant.class);
 	}
 
 	@Override
@@ -296,19 +315,6 @@ public final class CompatProviderImpl implements CompatProvider {
 	}
 
 	@Override
-	public @Nullable String getItemSNBT(@ReadOnly ItemStack itemStack) {
-		Validate.notNull(itemStack, "itemStack is null");
-		assert itemStack != null;
-		if (ItemUtils.isEmpty(itemStack)) {
-			return null;
-		}
-
-		var nmsItem = this.asNMSItemStack(itemStack);
-		var itemTag = this.getItemStackTag(nmsItem);
-		return itemTag.toString();
-	}
-
-	@Override
 	public ItemStackMetaTag getItemStackMetaTag(@ReadOnly @Nullable ItemStack itemStack) {
 		if (ItemUtils.isEmpty(itemStack)) {
 			return new ItemStackMetaTag(null);
@@ -392,7 +398,7 @@ public final class CompatProviderImpl implements CompatProvider {
 			itemTag.put("components", componentsTag);
 		}
 
-		var currentDataVersion = Bukkit.getUnsafe().getDataVersion();
+		var currentDataVersion = ServerUtils.getDataVersion();
 		var convertedItemTag = (CompoundTag) DataFixers.getDataFixer().update(
 				References.ITEM_STACK,
 				new Dynamic<>(Unsafe.castNonNull(NbtOps.INSTANCE), itemTag),
@@ -411,119 +417,11 @@ public final class CompatProviderImpl implements CompatProvider {
 		return Unsafe.assertNonNull(CraftItemStack.asCraftMirror(nmsItem));
 	}
 
-	// MC 1.21+ TODO Can be removed once we only support Bukkit 1.21+
-
 	@Override
-	public boolean isDestroyingBlocks(EntityExplodeEvent event) {
-		return isDestroyingBlocks(event.getExplosionResult());
-	}
-
-	@Override
-	public boolean isDestroyingBlocks(BlockExplodeEvent event) {
-		return isDestroyingBlocks(event.getExplosionResult());
-	}
-
-	private static boolean isDestroyingBlocks(ExplosionResult explosionResult) {
-		return explosionResult == ExplosionResult.DESTROY
-				|| explosionResult == ExplosionResult.DESTROY_WITH_DECAY;
-	}
-
-	// MC 1.21.3+ TODO Can be removed once we only support Bukkit 1.21.3+
-
-	@Override
-	public void setSalmonVariant(Salmon salmon, String variant) {
-		Salmon.Variant variantValue = EnumUtils.valueOf(Salmon.Variant.class, variant);
-		if (variantValue == null) {
-			variantValue = Salmon.Variant.MEDIUM; // Default
-		}
-		salmon.setVariant(variantValue);
-	}
-
-	// MC 1.21.5+ TODO Can be removed once we only support Bukkit 1.21.5+
-	// Actually, Paper differs in how registries are accessed.
-
-	// Paper-specific
-	@Override
-	public void setCowVariant(Cow cow, NamespacedKey variant) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.COW_VARIANT);
-		Cow.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Cow.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-
-		try {
-			cowSetVariantMethod.invoke(cow, variantValue);
-		} catch (Exception e) {
-			// Unexpected:
-			Log.severe("Failed to set cow variant!", e);
-		}
-	}
-
-	// Paper-specific
-	@Override
-	public NamespacedKey cycleCowVariant(NamespacedKey variant, boolean backwards) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.COW_VARIANT);
-		Cow.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Cow.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-
-		Registry<Cow.Variant> registryNonNull = Unsafe.castNonNull(registry);
-		return RegistryUtils.cycleKeyed(registryNonNull, variantValue, backwards).getKey();
-	}
-
-	// Paper-specific
-	@Override
-	public void setPigVariant(Pig pig, NamespacedKey variant) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.PIG_VARIANT);
-		Pig.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Pig.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-		pig.setVariant(variantValue);
-	}
-
-	// Paper-specific
-	@Override
-	public NamespacedKey cyclePigVariant(NamespacedKey variant, boolean backwards) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.PIG_VARIANT);
-		Pig.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Pig.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-
-		Registry<Pig.Variant> registryNonNull = Unsafe.castNonNull(registry);
-		return RegistryUtils.cycleKeyed(registryNonNull, variantValue, backwards).getKey();
-	}
-
-	// Paper-specific
-	@Override
-	public void setChickenVariant(Chicken chicken, NamespacedKey variant) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.CHICKEN_VARIANT);
-		Chicken.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Chicken.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-		chicken.setVariant(variantValue);
-	}
-
-	// Paper-specific
-	@Override
-	public NamespacedKey cycleChickenVariant(NamespacedKey variant, boolean backwards) {
-		var registry = RegistryAccess.registryAccess().getRegistry(RegistryKey.CHICKEN_VARIANT);
-		Chicken.Variant variantValue = registry.get(variant);
-		if (variantValue == null) {
-			variantValue = Chicken.Variant.TEMPERATE; // Default
-		}
-		assert variantValue != null;
-
-		Registry<Chicken.Variant> registryNonNull = Unsafe.castNonNull(registry);
-		return RegistryUtils.cycleKeyed(registryNonNull, variantValue, backwards).getKey();
+	public <T extends Keyed> Registry<T> getRegistry(Class<T> clazz) {
+		// Non-null: Expected to only be used with known registry types.
+		RegistryKey<T> registryKey = Unsafe.castNonNull(CLASS_TO_REGISTRY_KEY.get(clazz));
+		return Unsafe.castNonNull(RegistryAccess.registryAccess().getRegistry(registryKey));
 	}
 
 	// MC 1.21.9+ TODO Can be removed once we only support Bukkit 1.21.9+